# Support performance timings of heavy operations