--- a/src/scoped_timer.h
+++ b/src/scoped_timer.h
@@ -1,6 +1,7 @@
 #pragma once
 
-class ScopedTimer {
+class ScopedTimer
+{
 public:
   ScopedTimer(const std::string& name) : name(name) {
     t.start();