Now that there are tests being done comparing our CGAL (corefinement) rendering with alternatives (e.g. [against exciting new Manifold library](https://elalish.blogspot.com/2022/03/manifold-performance.html)), it's time to look at the details of the timings :-)

(reused code from https://github.com/openscad/openscad/pull/3641)

To build for performance comparisons
```bash
git clone https://github.com/ochafik/openscad.git --single-branch -b perf-timings openscad-perf-timings
cd openscad-perf-timings
git submodule init && git submodule update

mkdir build && cd build && cmake .. \
  -DCMAKE_BUILD_TYPE=Release -DEXPERIMENTAL=ON -DPERFORMANCE_TIMINGS=ON
```

I tried to match the testing conditions mentioned by the author of the [Manifold library](https://github.com/elalish/manifold) in [his post](https://elalish.blogspot.com/2022/03/manifold-performance.html):

```js
$fn=320; // Gives ~101k faces per sphere
// $fn=1000 // Gives ~1M faces per sphere
// $fn=3200 // Gives ~10M faces per sphere

difference() {
  sphere();
  translate([0.5, 0, 0]) sphere();
}
```

Tested with:
```bash
# MacOS:
# ln -s ./build/OpenSCAD.app/Contents/MacOS/OpenSCAD openscad
 
time ./openscad bigspheres.scad --enable=fast-csg --enable=fast-csg-trust-corefinement \
    -o diff.off '-D$fn=1000'

# STL is just so much slower
```

Results:

| $fn | # faces in each sphere | Total time (STL) | Total time (OFF) | PolySet -> hybrid | Corefinement Difference time |
| -- | -- | -- | -- | -- | -- |
| 100 | ~10k | 0.3sec | 0.26sec | 0.01sec | 0.08sec |
| 320 | ~101k | 1.5sec | 0.9sec | 0.06sec | 0.35sec |
| 1000 | ~1M | 13.9sec | 7.5sec | 0.63sec | 3.3sec |
| 3200 | ~10M | 173sec | 87sec | 7.7sec | 38.7sec |

I need to instrument imports & exports to understand where the most time is spent, but it seems the underlying union time in corefinement is more comparable to the (*mono-threaded CPU*) figures published for Manifold at the 100k and 1M faces.

@elalish could you share your raw figures and methodology?