@ochafik 2022-03-07
@sloriot FYI first benchmarks of your corefinement routines against fresh new [Manifold](https://github.com/elalish/manifold) library (see [post](https://elalish.blogspot.com/2022/03/manifold-performance.html)).