@ochafik 2022-03-08
Also, ported one of Manifold's microbenchmarks to CGAL in https://github.com/elalish/manifold/pull/84