@elalish 2022-03-08
@ochafik Looks like you've made a much better comparison than I did. I was running OpenSCAD's GUI, did an F5 render to warm things up, then F6 and recorded the "total rendering time" it conveniently displays. Of course I don't know exactly what this means, but considering the large difference between F5 and F6, I knew it didn't contain export and import was probably pretty trivial. Anyway, here's my [spreadsheet](https://docs.google.com/spreadsheets/d/11Vj3zJnoQ7W5-6SNOv7lIL36bJebtqDldJjIvTb7GxU/edit?usp=sharing).

If you want to benchmark the sponge too, I used this SCAD code:
```
difference(){
    cube(center=true);
    holes();
    rotate([90,0,0])holes();
    rotate([0,90,0])holes();
}

module holes()
import("mengerHoles3.stl");
```
And I exported "mengerHolesX.stl" from this line, to avoid a bunch of lazy unions that last I checked were still not lazy in OpenSCAD: https://github.com/elalish/manifold/blob/master/samples/src/menger_sponge.cpp#L50

How do these libraries compare performance-wise on your machine?