@sloriot 2022-03-08
Hey, thanks for the notification. Very interesting. You are using EPECK as input kernel while Manifold uses double, right? I wonder what difference it makes if we are also using double as input with corefinement. Also do you know if Manifold is taking advantage of several threads?