@ochafik 2022-03-08
@sloriot I’ve tested Epick and Epeck against manifold in a very limited microbenchmark here, Manifold monothreaded cpu mode seems >2x faster than coref on epick for Union: https://github.com/elalish/manifold/pull/84

(Not sure corefinement would accept a simpler ‘Cartesian<double>’?)

Manifold can leverage CUDA and OpenMP but I haven’t tested either, stuck to mono-threaded CPU comparisons for now (unfair to Manifold haha) but @elalish seems to say we can expect at least 2x speedups, even before he proceeds to optimizing more steps in his library.

Could corefinement use TBB to parallelize some operations btw? Looks like Epeck is rather thread safe from reading https://github.com/CGAL/cgal/issues/6141.