@sloriot 2022-03-08
We could indeed use TBB to process intersections and retriangulation but I never spent time on it as no one asked for it. I saw the bench. It is strange that difference is faster but not union. One possibility could be that union need to import more triangles than difference. If that the case, then we might be impacted by the generic import rather than using something dedicated to `Surface_mesh`.