@ochafik 2022-03-08
> running OpenSCAD's GUI, did an F5 render to warm things up, then F6 and recorded the "total rendering time" it conveniently displays. Of course I don't know exactly what this means, but considering the large difference between F5 and F6, I knew it didn't contain export and import was probably pretty trivial

@elalish This should indeed skip import/export, but may still have some conversions I've left to avoid reimplementing rendering logic for fast-csg's hybrid polyhedron's data structures (which can be either Surface_mesh or a differently-kernelled Nef than the one already supported in the renderer). Given how big the meshes are, even just copies to the cache might take time here.